% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.R
\name{generic}
\alias{generic}
\alias{bookdown}
\alias{distill}
\title{Codefolding}
\usage{
generic(query, style)

bookdown(query = "pre.sourceCode",
  style = "position: absolute; right: 5\%; z-index: 200")

distill(query = "d-code", style = "")
}
\arguments{
\item{query}{\code{character} CSS-selector string for code blocks}

\item{style}{\code{character} CSS style for the button}
}
\value{
\code{htmltools::tagList}
}
\description{
Codefolding
}
